фыафыа

+++ TerraZ_Client/IndexTypes.cs
@@ -1,4 +1,4 @@
-﻿public enum IndexTypes : int
+﻿public enum IndexTypes : byte
 {
     Authorization = 1,
     //PlayerShifting = 2,
